refactor: Decouple x-ssr from x-each
