+++ src/directives/each.ts
@@ -1,4 +1,4 @@
-import getParents from "../get-parents";
+import { getLevel, getTemplates } from "../utils";
 import type { DirectiveParameters, ExtendedHTMLElement } from "../types";
 
 function eachDirective({
@@ -10,7 +10,7 @@ function eachDirective({
   directives,
 }: DirectiveParameters) {
   const elementState = getState(element);
-  let state = evaluate(expression, elementState) || [];
+  const state = evaluate(expression, elementState) || [];
 
   if (!Array.isArray(state)) {
     console.error(
@@ -26,9 +26,7 @@ function eachDirective({
     return;
   }
 
-  const level =
-    getParents(element, "x-recurse").length +
-    (element.hasAttribute("x-recurse") ? 1 : 0);
+  const level = getLevel(element);
   const amountOfTemplates = element.templates?.length || 0;
 
   if (amountOfTemplates > 0) {
@@ -84,8 +82,7 @@ function eachDirective({
     // In case state is derived from another x-each, use getState(element) here.
     element.state = hasParentEach ? elementState.state : state;
 
-    const xTemplates = getTemplates(element);
-    let templates = xTemplates;
+    let templates = getTemplates(element);
 
     if (!templates.length) {
       console.error("x-each - x-template was not found", element);
@@ -106,27 +103,6 @@ function eachDirective({
     element.setAttribute("_x-init", "1");
     element.setAttribute("x-has-each", "");
 
-    if (element.hasAttribute("x-ssr")) {
-      element.state = extractValuesFromTemplates(element, xTemplates, level);
-
-      // Find the closest state container and update its internal state
-      const parentStateElement: ExtendedHTMLElement | null =
-        element.closest("[x-state]");
-      if (parentStateElement) {
-        const parentKey = expression.split("state.")[1];
-        parentStateElement.state = {
-          ...parentStateElement.state,
-          [parentKey]: element.state,
-        };
-      }
-
-      // Now that the state has been captured, go back to normal flow
-      element.removeAttribute("x-ssr");
-
-      // Use the state now
-      state = element.state;
-    }
-
     // Empty contents as they'll be replaced by applying the template
     while (element.firstChild) {
       element.lastChild && element.removeChild(element.lastChild);
@@ -140,102 +116,6 @@ function eachDirective({
   evaluateDirectives(directives, element);
 }
 eachDirective.evaluateFrom = "top";
-
-function getTemplates(element: ExtendedHTMLElement) {
-  return element.querySelectorAll(":scope > *[x-template]");
-}
-
-function extractValuesFromTemplates(
-  element: ExtendedHTMLElement,
-  xTemplates: ExtendedHTMLElement["templates"],
-  level: number
-) {
-  const ret = [];
-
-  if (!xTemplates) {
-    return [];
-  }
-
-  for (let i = 0; i < xTemplates.length; i++) {
-    const xTemplate = xTemplates[i] as ExtendedHTMLElement;
-    const newState = getValues(element, xTemplate);
-    const xEachContainers = xTemplate.querySelectorAll("[x-each]");
-
-    // The element should be a state container itself so that children can
-    // access its data.
-    xTemplate.setAttribute("x-state", "");
-
-    for (let j = 0; j < xEachContainers.length; j++) {
-      const xEachContainer = xEachContainers[j] as ExtendedHTMLElement;
-
-      // Pick only elements that have the x-each as their parent.
-      // The gotcha here is that since the element itself is x-each,
-      // closest() matches to itself so you should traverse starting from
-      // the immediate parent.
-      if (xEachContainer.parentElement?.closest("[x-each]") === element) {
-        const xEachValue = xEachContainer.getAttribute("x-each") || "";
-        const k = xEachValue.split("state.value.")[1];
-
-        const v = extractValuesFromTemplates(
-          xEachContainer,
-          getTemplates(xEachContainer),
-          level + 1
-        );
-
-        if (k) {
-          // TODO: Can unpacking be avoided for arrays?
-          if (Array.isArray(v)) {
-            // @ts-ignore How to type this?
-            newState[k] = v.map((i) => i[0]);
-          } else {
-            // @ts-ignore How to type this?
-            newState[k] = v;
-          }
-
-          // The actual state is stored to the object
-          xTemplate.state = { value: v, level };
-        }
-      }
-    }
-
-    ret.push(newState);
-  }
-
-  return ret;
-}
-
-function getValues(
-  element: ExtendedHTMLElement,
-  xTemplate: ExtendedHTMLElement
-) {
-  const xValues = xTemplate.hasAttribute("x")
-    ? [xTemplate]
-    : xTemplate.querySelectorAll("[x]");
-  const newObjectState: Record<string, unknown> = {};
-  const newArrayState = [];
-
-  for (let j = 0; j < xValues.length; j++) {
-    const xValue = xValues[j];
-
-    // Pick only elements that have the x-each as their parent
-    if (xValue.closest("[x-each]") === element) {
-      const xProperty = xValue.getAttribute("x") || "";
-      const v = xValue.textContent;
-
-      if (xProperty === "state.value") {
-        newArrayState.push(v);
-      } else {
-        const k = xProperty.split("state.value.")[1];
-
-        if (k) {
-          newObjectState[k] = v;
-        }
-      }
-    }
-  }
-
-  return newArrayState.length > 0 ? newArrayState : newObjectState;
-}
 
 function getTemplateRenderer(
   element: ExtendedHTMLElement,
